reimplemented post_likes many to many relationship with extra entity using 1:n - n:1 relations, implemented repository class for this new entity, modified like logic to check first if its already liked or not. Also removed orphanRemoval because of unwanted deletes on complicated associations between entities

+++ src/main/java/com/pse/testserver/repository/EventRepository.java
@@ -48,7 +48,7 @@ public interface EventRepository extends CrudRepository<Event, Integer> {
      * @return list of all events containing the string name with the specified category
      */
     @Query(value = "Select * from events e where e.name like %:name% and e.category_id like :category", nativeQuery = true)
-    List<Event> findAllByNameSortedByCategory(@Param("name") String name, @Param("category") long category);
+    List<Event> findAllByNameSortedByCategory(@Param("name") String name, @Param("category") int category);
 
 
     /**
